requirements-core.txt
numpy
pandas
scikit-learn
matplotlib
pyyaml
pillow==11.2.1
# federated / утилиты
flwr[simulation]==1.22.0
flwr-datasets[vision]>=0.5.0
torchmetrics==1.7.1
rich
tqdm